tar/: add "platform.h" to all .c files
This fixes the client being broken on 32-bit platforms (i386 and ARM).

The problem is that I was unaware that glibc processed the "feature test
macros" [1] the first time the compiler encountered any #include
statements for system libraries [2].  As a result, when I removed the
"unnecessary" bsdtar_platform.h includes [3], that meant that some
object files interpreted off_t and ino_t as being 4 bytes, while other
object files interpreted those variables as being 8 bytes.  That led to
all sorts of Bad Things (tm) happening.

Fortunately, one of those Bad Things (tm) was that the networking code
could not read the machine number correctly, so the client could not
communicate with the server.  No user data was at risk from anybody
using the unreleased git master.

[1] https://www.gnu.org/software/libc/manual/html_node/Feature-Test-Macros.html
[2] I tested extensively on amd64, as did Travis-CI, but I neglected to
    test on a 32-bit system.
[3] remove-unnecessary-bsdtar-h branch: 1f56dc7c35b32d9ce37d727215c1293e59d0f2e1